+++ CZJShop/ThirdSDK/EasyJSWebView/EasyJSWebViewProxyDelegate.m
@@ -133,7 +133,7 @@ return EasyJS.call(obj, method, Array.prototype.slice.call(arguments));\
 									  stringByReplacingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
 			
 			NSArray* formattedArgs = [argsAsString componentsSeparatedByString:@":"];
-			for (int i = 0, j = 0, l = [formattedArgs count]; i < l; i+=2, j++){
+			for (NSInteger i = 0, j = 0, l = [formattedArgs count]; i < l; i+=2, j++){
 				NSString* type = ((NSString*) [formattedArgs objectAtIndex:i]);
 				NSString* argStr = ((NSString*) [formattedArgs objectAtIndex:i + 1]);
